Server/requirements.txt
transformers==4.46.3
torchaudio==2.5.1
Flask==3.1.0
gunicorn==23.0.0
python-Levenshtein==0.26.1

# transformers>=4.44.1 
# torchaudio>=2.4.1
# Flask>=3.0.0
# gunicorn>=23.0.0
# python-Levenshtein